requirements.txt
python-telegram-bot==20.7
pymongo==4.6.1
google-generativeai==0.4.0
requests==2.31.0
python-dotenv==1.0.1
